+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v2.0.0-rc.7
 ----
  * ons-switch: Trigger `change` event only once.
  * ons-bottom-toolbar: Correctly register element. Fix [#1426](https://github.com/OnsenUI/OnsenUI/issues/1426).
+ * ons-pull-hook: Fix [#1415](https://github.com/OnsenUI/OnsenUI/issues/1415).
 
 v2.0.0-rc.6
 ----